Move opts to daemon/pkg/opts
