+++ daemon/cluster/swarm.go
@@ -13,8 +13,8 @@ import (
 	"github.com/docker/docker/api/types/filters"
 	types "github.com/docker/docker/api/types/swarm"
 	"github.com/docker/docker/daemon/cluster/convert"
+	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/opts"
 	"github.com/docker/docker/pkg/stack"
 	swarmapi "github.com/moby/swarmkit/v2/api"
 	"github.com/moby/swarmkit/v2/manager/encryption"
