+++ daemon/command/daemon.go
@@ -30,6 +30,7 @@ import (
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/internal/libcontainerd/supervisor"
 	"github.com/docker/docker/daemon/listeners"
+	dopts "github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/daemon/pkg/plugin"
 	apiserver "github.com/docker/docker/daemon/server"
 	"github.com/docker/docker/daemon/server/middleware"
@@ -49,7 +50,6 @@ import (
 	"github.com/docker/docker/daemon/server/router/volume"
 	"github.com/docker/docker/dockerversion"
 	"github.com/docker/docker/internal/otelutil"
-	dopts "github.com/docker/docker/opts"
 	"github.com/docker/docker/pkg/authorization"
 	"github.com/docker/docker/pkg/homedir"
 	"github.com/docker/docker/pkg/pidfile"
