+++ daemon/config/config_linux.go
@@ -12,8 +12,8 @@ import (
 	"github.com/containerd/log"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/system"
+	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/libnetwork/drivers/bridge"
-	"github.com/docker/docker/opts"
 	"github.com/docker/docker/pkg/homedir"
 	"github.com/docker/docker/pkg/rootless"
 	"github.com/pkg/errors"
