+++ daemon/config/config_windows_test.go
@@ -3,7 +3,7 @@ package config
 import (
 	"testing"
 
-	"github.com/docker/docker/opts"
+	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/spf13/pflag"
 	"gotest.tools/v3/assert"
 	is "gotest.tools/v3/assert/cmp"
