+++ daemon/container.go
@@ -16,10 +16,10 @@ import (
 	"github.com/docker/docker/daemon/config"
 	"github.com/docker/docker/daemon/container"
 	"github.com/docker/docker/daemon/network"
+	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/oci/caps"
-	"github.com/docker/docker/opts"
 	volumemounts "github.com/docker/docker/volume/mounts"
 	"github.com/docker/go-connections/nat"
 	"github.com/moby/sys/signal"
