+++ daemon/daemon_unix.go
@@ -31,6 +31,7 @@ import (
 	"github.com/docker/docker/daemon/initlayer"
 	"github.com/docker/docker/daemon/internal/libcontainerd/remote"
 	"github.com/docker/docker/daemon/internal/usergroup"
+	"github.com/docker/docker/daemon/pkg/opts"
 	"github.com/docker/docker/errdefs"
 	"github.com/docker/docker/internal/nlwrap"
 	"github.com/docker/docker/internal/otelutil"
@@ -40,7 +41,6 @@ import (
 	"github.com/docker/docker/libnetwork/netlabel"
 	"github.com/docker/docker/libnetwork/options"
 	lntypes "github.com/docker/docker/libnetwork/types"
-	"github.com/docker/docker/opts"
 	"github.com/docker/docker/pkg/sysinfo"
 	"github.com/docker/docker/runconfig"
 	volumemounts "github.com/docker/docker/volume/mounts"
