+++ daemon/pkg/opts/hosts.go
