+++ daemon/pkg/opts/hosts_test.go
