+++ daemon/pkg/opts/hosts_windows.go
